# lz4block: Use unsigned indexes