This allows the compiler to simplify some bounds checks.

Benchmark results on amd64 with -tags noasm:

```
name                old time/op    new time/op    delta
Uncompress-8           499ns ± 3%     511ns ± 2%  +2.42%  (p=0.000 n=20+20)
UncompressPg1661-8    1.13ms ± 1%    1.10ms ± 1%  -2.20%  (p=0.000 n=20+20)
UncompressDigits-8    77.4µs ± 1%    75.9µs ± 1%  -1.92%  (p=0.000 n=20+20)
UncompressTwain-8      724µs ± 1%     702µs ± 1%  -3.02%  (p=0.000 n=20+19)
UncompressRand-8      3.67µs ± 2%    3.61µs ± 2%  -1.55%  (p=0.000 n=20+20)

name                old speed      new speed      delta
UncompressPg1661-8   527MB/s ± 1%   539MB/s ± 1%  +2.25%  (p=0.000 n=20+20)
UncompressDigits-8  1.29GB/s ± 1%  1.32GB/s ± 1%  +1.96%  (p=0.000 n=20+20)
UncompressTwain-8    536MB/s ± 1%   553MB/s ± 1%  +3.11%  (p=0.000 n=20+19)
UncompressRand-8    4.46GB/s ± 2%  4.53GB/s ± 2%  +1.57%  (p=0.000 n=20+20)
```

(Allocation rates are all the same.)

Btw., it looks like BenchmarkUncompress is broken. It calls UncompressBlock on an LZ4 stream, but never checks the error.